Fix PX4_GZ_MODEL

+++ entrypoint.sh
@@ -17,10 +17,6 @@ if [[ "${START_SSHD:-1}" == "1" ]]; then
     /usr/sbin/sshd -f /etc/ssh/sshd_config_test_clion
 fi
 
-if [[ -d /root/PX4-Autopilot/Tools/simulation/gz ]]; then
-    patch_px4_gz_models
-fi
-
 if [[ "${RUN_PX4_SETUP:-1}" == "1" ]]; then
     if [[ -f /root/PX4-Autopilot/Tools/setup/ubuntu.sh ]]; then
         setup_px4_deps
@@ -32,16 +28,20 @@ fi
 
 if [[ "${BUILD_OPENVINS:-1}" == "1" ]]; then
     if [[ -d /root/colcon_ws/src ]] && find /root/colcon_ws/src -name package.xml -print -quit | grep -q .; then
-        apt-get update
+        if [[ -f /root/colcon_ws/install/setup.bash && "${FORCE_BUILD_OPENVINS:-0}" != "1" ]]; then
+            echo "OpenVINS build skipped: /root/colcon_ws/install/setup.bash already exists"
+        else
+            apt-get update
 
-        cd /root/colcon_ws
-        rosdep install --from-paths src --ignore-src -r -y
+            cd /root/colcon_ws
+            rosdep install --from-paths src --ignore-src -r -y
 
-        MAKEFLAGS="${MAKEFLAGS:-"-j2"}" colcon build --symlink-install \
-            --executor sequential \
-            --packages-select ov_core ov_init ov_msckf ov_eval
+            MAKEFLAGS="${MAKEFLAGS:-"-j2"}" colcon build --symlink-install \
+                --executor sequential \
+                --packages-select ov_core ov_init ov_msckf ov_eval
 
-        rm -rf /var/lib/apt/lists/*
+            rm -rf /var/lib/apt/lists/*
+        fi
     else
         echo "OpenVINS build skipped: no ROS packages found under /root/colcon_ws/src"
     fi
